wip refactoring: cleanup
